+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - update TODO's
 - update main workflow
+- add basic router
+- add home and error views
 
 ## 0.0.1